[REFACTOR] transferido a abstração de dados do usuário logado para o Core de api
