+++ src/building blocks/Estudos.NSE.WebApi.Core/Usuario/ClaimsPrincipalExtensions.cs
@@ -1,0 +1,41 @@
+﻿using System;
+using System.Security.Claims;
+
+namespace Estudos.NSE.WebApi.Core.Usuario
+{
+    public static class ClaimsPrincipalExtensions
+    {
+        public static string GetUserId(this ClaimsPrincipal principal)
+        {
+            if (principal == null)
+            {
+                throw new ArgumentException(nameof(principal));
+            }
+
+            var claim = principal.FindFirst("sub");
+            return claim?.Value;
+        }
+
+        public static string GetUserEmail(this ClaimsPrincipal principal)
+        {
+            if (principal == null)
+            {
+                throw new ArgumentException(nameof(principal));
+            }
+
+            var claim = principal.FindFirst("email");
+            return claim?.Value;
+        }
+
+        public static string GetUserToken(this ClaimsPrincipal principal)
+        {
+            if (principal == null)
+            {
+                throw new ArgumentException(nameof(principal));
+            }
+
+            var claim = principal.FindFirst("JWT");
+            return claim?.Value;
+        }
+    }
+}
