+++ src/building blocks/Estudos.NSE.WebApi.Core/Usuario/IAspNetUser.cs
@@ -1,0 +1,20 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Security.Claims;
+using Microsoft.AspNetCore.Http;
+
+namespace Estudos.NSE.WebApi.Core.Usuario
+{
+    public interface IAspNetUser
+    {
+        string Name { get; }
+        Guid ObterUserId();
+        string ObterUserEmail();
+        string ObterUserToken();
+        bool EstaAutenticado();
+        bool PossuiRole(string role);
+        public bool PossuiRole(string[] roles);
+        IEnumerable<Claim> ObterClaims();
+        HttpContext ObterHttpContext();
+    }
+}
