+++ src/web/Estudos.NSE.WebApp.MVC/Estudos.NSE.WebApp.MVC.csproj
@@ -17,6 +17,7 @@
 
   <ItemGroup>
     <ProjectReference Include="..\..\building blocks\Estudos.NSE.Core\Estudos.NSE.Core.csproj" />
+    <ProjectReference Include="..\..\building blocks\Estudos.NSE.WebApi.Core\Estudos.NSE.WebApi.Core.csproj" />
   </ItemGroup>
 
 </Project>
