+++ src/web/Estudos.NSE.WebApp.MVC/Views/Shared/_LoginPartial.cshtml
@@ -1,5 +1,6 @@
 ﻿@using Estudos.NSE.WebApp.MVC.Extensions
-@inject IUser AppUser
+@using Estudos.NSE.WebApi.Core.Usuario
+@inject IAspNetUser AppUser
 
 @if (AppUser.EstaAutenticado())
 {
